refactor(plugin): makes several functions reusable accross the plugins pkg
This patch ensures that we can reuse several functions needed for the
hibernation command avoiding the creation of duplicated code.

Partially closes #781
